@@ -36,23 +36,24 @@
       "gulp"
    ],
    "dependencies": {
+      "@angular/cli": "^6.0.7",
       "angular2-virtual-scroll": "0.3.1",
       "lodash": "4.17.10",
       "reflect-metadata": "0.1.10",
       "tslib": "1.7.1"
    },
    "devDependencies": {
-      "@angular/animations": "4.4.3",
-      "@angular/cli": "1.6.5",
-      "@angular/common": "4.4.3",
-      "@angular/compiler": "4.4.3",
-      "@angular/compiler-cli": "4.4.3",
-      "@angular/core": "4.4.3",
-      "@angular/forms": "4.4.3",
-      "@angular/http": "4.4.3",
-      "@angular/platform-browser": "4.4.3",
-      "@angular/platform-browser-dynamic": "4.4.3",
-      "@angular/router": "4.4.3",
+      "@angular/animations": "6.0.3",
+      "@angular/common": "6.0.3",
+      "@angular/compiler": "6.0.3",
+      "@angular/compiler-cli": "6.0.3",
+      "@angular/core": "6.0.3",
+      "@angular/forms": "6.0.3",
+      "@angular/http": "6.0.3",
+      "@angular/platform-browser": "6.0.3",
+      "@angular/platform-browser-dynamic": "6.0.3",
+      "@angular/router": "6.0.3",
+      "@angular/tsc-wrapped": "4.4.6",
       "@stratio/egeo-ui-base": "1.4.1",
       "@types/chalk": "0.4.31",
       "@types/fs-extra": "3.0.3",
@@ -102,7 +103,7 @@
       "rollup-plugin-node-resolve": "3.0.0",
       "rollup-plugin-strip-banner": "0.2.0",
       "run-sequence": "2.2.0",
-      "rxjs": "5.5.2",
+      "rxjs": "6.2.0",
       "scss-bundle": "2.1.0",
       "showdown": "1.8.2",
       "sorcery": "0.10.0",
@@ -110,10 +111,11 @@
       "ts-node": "3.2.0",
       "tsconfig-paths": "2.2.0",
       "tslint": "5.5.0",
-      "typescript": "2.3.4",
+      "typescript": "2.7.2",
       "uglify-js": "2.8.22",
       "word-wrap": "1.2.3",
-      "zone.js": "0.8.18"
+      "zone.js": "0.8.26",
+      "@angular-devkit/build-angular": "~0.6.6"
    },
    "config": {
       "commitizen": {
