@@ -13,7 +13,7 @@ import {
    OnChanges, OnDestroy, OnInit, ViewChildren, Output, EventEmitter
 } from '@angular/core';
 import { ControlValueAccessor, FormControl, NG_VALIDATORS, NG_VALUE_ACCESSOR } from '@angular/forms';
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription } from 'rxjs';
 
 import { StTextareaError } from './st-textarea.error.model';
 
